pyproject.toml
[project]
name = "flatlandian"
version = "0.1.4"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pdoc>=16.0.0",
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]

[tool.mypy]
strict = true
disallow_any_unimported = true
no_implicit_optional = true
warn_unreachable = true
show_error_codes = true
pretty = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",  # missing-trailing-comma: conflicts with ruff formatter
    "D105",  # undocumented-magic-method
    "S311",  # suspicious-non-cryptographic-random-usage
]

# Allow 'unsafe' fixes:
extend-safe-fixes = [
    "TC",  # flake8-type-checking
]

[tool.ruff.lint.per-file-ignores]
# Not relevant in tests:
"tests/*" = [
    "D103",  # undocumented-public-function
    "S101",  # assert
    "PLR2004",  # magic-value-comparison
]

[tool.ruff.lint.pydocstyle]
convention = "google"
